+++ src/components/Table/TableBody.tsx
@@ -1,6 +1,14 @@
 import { cn } from "../../utils/cn";
 import type { TableBodyProps } from "./Table.types";
 
-export function TableBody({ children, className }: TableBodyProps) {
-  return <tbody data-react-fancy-table-body="" className={cn("divide-y divide-zinc-100 dark:divide-zinc-800", className)}>{children}</tbody>;
+export function TableBody({ children, className, ...rest }: TableBodyProps) {
+  return (
+    <tbody
+      {...rest}
+      data-react-fancy-table-body=""
+      className={cn("divide-y divide-zinc-100 dark:divide-zinc-800", className)}
+    >
+      {children}
+    </tbody>
+  );
 }